fix(UserServiceImpl): prevent saving users with duplicate emails. Previously, duplicate email exception was thrown. However,their data was saved to the database.This fix ensures that the data with duplicate emails  are not saved to the db.

+++ src/main/java/com/example/demo/book/BookServiceImpl.java
@@ -41,8 +41,12 @@ public class BookServiceImpl implements BookService{
     }
 
     @Override
-    public void deleteBook(int isbn) {
-        bookRepo.deleteById(isbn);
+    public Boolean deleteBook(int isbn) {
+        if(bookRepo.existsByIsbn(isbn)) {
+            bookRepo.deleteById(isbn);
+            return true;
+        } else
+            return false;
     }
 
     @Override
